Move `has_comments` to `CommentRanges`
