+++ crates/ruff_linter/src/rules/pylint/rules/nested_min_max.rs
@@ -155,7 +155,11 @@ pub(crate) fn nested_min_max(
         MinMax::try_from_call(func.as_ref(), keywords.as_ref(), checker.semantic()) == Some(min_max)
     }) {
         let mut diagnostic = Diagnostic::new(NestedMinMax { func: min_max }, expr.range());
-        if !checker.indexer().has_comments(expr, checker.locator()) {
+        if !checker
+            .indexer()
+            .comment_ranges()
+            .has_comments(expr, checker.locator())
+        {
             let flattened_expr = Expr::Call(ast::ExprCall {
                 func: Box::new(func.clone()),
                 arguments: Arguments {
